refactor(fft): merge packed-Stockham forward and inverse modules
Consolidate the packed-Stockham FFT (R2C) and its inverse (C2R) into a
single fftPackedStockham/ module: shared code at the top level, with the
forward- and inverse-specific code split into r2c/ and c2r/ subfolders.

- Unify the mixed-radix butterfly math into shared combineRadixN helpers
  parametrised by a twiddleSign override (folded at pipeline creation),
  making it the single source across transform, multiPassStage and
  transformInPlaceMixed for both directions.
- Share the params ring, trig-table helpers and dispatch runtime.

No behaviour change; fourier and batchRange GPU tests stay green (2760).
