+++ packages/fft/src/fourier/__test__/benchHarness.ts
@@ -1,8 +1,8 @@
 import { createBenchTimestamp, defaultBenchStatsConfig } from '@musetric/utils';
 import { createGpuContext } from '@musetric/utils/gpu';
 import { type FourierMode, type IFourierMode } from '../config.es.js';
+import { getPackedStockhamC2rVariant } from '../fftPackedStockham/c2r/support.js';
 import { fouriers } from '../fouriers.js';
-import { getPackedStockhamC2rVariant } from '../ifftPackedStockhamC2r/support.js';
 import { iffts } from '../iffts.js';
 import { type Fourier, type FourierBatchRange } from '../types.js';
 import {
