+++ packages/fft/src/fourier/fftPackedStockham/c2r/index.ts
@@ -1,0 +1,7 @@
+import { createStockhamFourier } from '../fourierRuntime.js';
+import { createStateCell } from './state.js';
+
+export const createIfftPackedStockhamC2r = createStockhamFourier(
+  createStateCell,
+  'packed-stockham-c2r-transform',
+);
