+++ packages/fft/src/fourier/fftPackedStockham/r2c/index.ts
@@ -1,0 +1,7 @@
+import { createStockhamFourier } from '../fourierRuntime.js';
+import { createStateCell } from './state.js';
+
+export const createFftPackedStockhamR2c = createStockhamFourier(
+  createStateCell,
+  'packed-stockham-r2c-transform',
+);
