+++ packages/fft/src/fourier/iffts.ts
@@ -1,5 +1,5 @@
 import { type IFourierMode } from './config.es.js';
-import { createIfftPackedStockhamC2r } from './ifftPackedStockhamC2r/index.js';
+import { createIfftPackedStockhamC2r } from './fftPackedStockham/c2r/index.js';
 import { type CreateFourier } from './types.js';
 
 export const iffts: Record<IFourierMode, CreateFourier> = {
