+++ packages/fft/src/index.ts
@@ -1,7 +1,7 @@
 export * from './fourier/cell.js';
-export * from './fourier/fftPackedStockhamR2c/index.js';
+export * from './fourier/fftPackedStockham/c2r/index.js';
+export * from './fourier/fftPackedStockham/r2c/index.js';
 export * from './fourier/fftPackedTiledR2c/index.js';
 export * from './fourier/fouriers.js';
-export * from './fourier/ifftPackedStockhamC2r/index.js';
 export * from './fourier/iffts.js';
 export * from './fourier/types.js';
